GET/{id} method in java done

+++ backend/src/main/java/org/factoriaf5/backend/repository/ReviewRepository.java
@@ -3,6 +3,6 @@ package org.factoriaf5.backend.repository;
 import org.factoriaf5.backend.models.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+
 }
